feat: 새로운 도구 및 에이전트 기능 추가, 코드 구조 개선
