+++ .pre-commit-config.yaml
@@ -11,7 +11,12 @@ repos:
     rev: 7.1.1
     hooks:
       - id: flake8
+        args:
+          - '--max-line-length=100'
+          - '--extend-ignore=E203'
   - repo: https://github.com/pylint-dev/pylint
     rev: v3.3.4
     hooks:
       - id: pylint
+        args:
+          - '--rcfile=.pylintrc'
