+++ src/agent/agent.py
@@ -1,5 +1,8 @@
 from .utils.state import AgentState
-from .utils.nodes import query_analysis, cc_agent, tool_node
+from .utils.constants import Category
+from .nodes.cc_agent import cc_agent
+from .nodes.query_analysis import query_analysis
+from .nodes.tool_node import tool_node
 from langgraph.graph import StateGraph, START, END
 from langgraph.graph.state import CompiledStateGraph
 from langchain_core.messages import AIMessage
@@ -13,16 +16,12 @@ def get_graph() -> CompiledStateGraph:
     builder.add_node("tools", tool_node)
 
     def query_condition(state: AgentState):
-        return END if state["is_end"] == True else "cc_agent"
+        return END if state["category"] == Category.PENDING_JUDGMENT.value else "cc_agent"
 
     def should_continue(state: AgentState):
         messages = state["messages"]
         last_message = messages[-1]
-        return (
-            "tools"
-            if isinstance(last_message, AIMessage) and last_message.tool_calls
-            else END
-        )
+        return "tools" if isinstance(last_message, AIMessage) and last_message.tool_calls else END
 
     builder.add_edge(START, "query_analysis")
     builder.add_conditional_edges("query_analysis", query_condition, [END, "cc_agent"])
