+++ src/agent/nodes/tool_node.py
@@ -1,0 +1,6 @@
+from ..tools.retriever_tool import retriever_tool
+from ..tools.express_carrier_tool import express_carrier_tool
+from langgraph.prebuilt import ToolNode
+
+tools = [retriever_tool, express_carrier_tool]
+tool_node = ToolNode(tools)
