+++ src/agent/utils/constants.py
@@ -1,4 +1,5 @@
 from getpass import getpass
+from enum import Enum
 from dotenv import load_dotenv
 import os
 
@@ -16,3 +17,18 @@ _set_env("MODEL")
 
 MODEL = os.environ["MODEL"]
 CSV_PATH = "./data/aIcs-ace.csv"
+
+
+class Category(str, Enum):
+    PERSONAL_INFORMATION = "개인정보"
+    TAX = "세금"
+    RELEASE = "출고"
+    PENDING = "미결"
+    INSPECTION = "검사"
+    ORIGIN = "원산지"
+    PROHIBITED = "금지"
+    MOVING_GOODS = "이사화물"
+    RELEASE_AGAIN = "출고"
+    CUSTOMS_CLEARANCE = "통관"
+    ETC = "기타"
+    PENDING_JUDGMENT = "판단 보류"
