+++ src/main.py
@@ -1,8 +1,8 @@
-from src.agent.agent import get_graph
-from src.agent.utils.visualize import save_graph_as_png
-from src.agent.utils.state import AgentState, initialize_state
-from src.agent.utils.logger import logger
-from langchain_core.messages import HumanMessage
+from .agent.agent import get_graph
+from .agent.utils.visualize import save_graph_as_png
+from .agent.utils.state import AgentState, initialize_state, update_state
+from .agent.utils.logger import logger
+from langchain_core.messages import AIMessage, HumanMessage
 from langchain.globals import set_debug, set_verbose
 from langchain_core.runnables import RunnableConfig
 from typing import Optional
@@ -14,21 +14,25 @@ set_verbose(True)
 
 def main():
     graph = get_graph()
-    config: Optional[RunnableConfig] = {"configurable": {"thread_id": "1"}}
     state: AgentState = initialize_state()
+    config: Optional[RunnableConfig] = {"configurable": {"thread_id": "1"}}
     save_graph_as_png(graph, "workflow.png")
 
     while True:
         try:
             question = input("질문을 입력하세요 (종료하려면 'q' 입력): ")
-            if question.lower() == "q":
+            if question.lower() == "q" or question.lower() == "quit":
                 break
+
             res = graph.invoke(
                 {"messages": [HumanMessage(content=question)]},
                 config,
                 stream_mode="values",
             )
-            print(f"답변: {res["messages"][-1].content}")
+
+            state = update_state(state, messages=res["messages"], category=res["category"])
+            if isinstance(res["messages"][-1], AIMessage):
+                print(f"답변: {res["messages"][-1].content}")
         except KeyboardInterrupt:
             print("\n프로그램을 종료합니다.")
             break
